refactor(query): IsAreaSafe を query へ移し公開を最小化する
IsAreaSafe は派閥と座標だけを読む純粋な安全判定で、休息・睡眠・読書など複数の
継続アクティビティが使う汎用の読み取り。activity で公開していたが query の charter に
沿う関数なので query/faction.go へ移し、query 慣例に合わせ world を先頭引数にする。
テストも faction_test.go へ移す。

sleep 専用で activity 内でしか使わない BeddingQualityAt は公開の必要がないので
beddingQualityAt へ戻す。当初これらを公開したのは EvaluateSleepConditions が
systems 側にあった名残で、activity へ集約した今は不要になった。
